refactored to match assignment
